fix(normalize): Drop invalid measurement names
